Extract the vault-format core into @glance-apps/obsidian-format
packages/obsidian-format/ now holds the line-format layer both dayGLANCE
and the bridge plugin consume — the same code, so two codebases can never
grow two lookalike parsers. THE PACKAGE IS FORMAT, NEVER POLICY: it should
not know an ownership rule exists. That sentence lives in its README and
its index header; the one seam where format brushes policy —
updateTaskLines' write-time title guard — is commented at the callsite:
report through onTitleConflict, never resolve here.

Contents (moved verbatim, byte-exact programmatic cuts): identity
(simpleHash, deriveBlockId + frozen goldens, legacyObsidianId,
splitBlockId, hasForeignBlockId, blockIdSuffix), completion markers, the
Tasks-metadata splitter, frontmatter, line building/rewriting/sorting
(parseLeadingTime, buildTimePrefix, stripLinePrefixes,
buildObsidianTaskLine, updateTaskLines, taskLineSortKey,
sortTaskLinesInSection), note naming (formatDatePattern, date parsers,
dailyNoteFilename, the portability validators), parseTasksFromMarkdown,
and the heartbeat shape. Staying in dayGLANCE: all four transports, both
sync functions, every 3.10 ownership rule, ghost containment, the deletion
detector, the write-policy gates, scan/window helpers.

Consumption: file:packages/obsidian-format from both dayGLANCE and the
plugin, under the final npm name from day one — published at the
extraction event (directory submission requires a root manifest, so the
plugin gets its own repo then) with zero import-specifier changes ever.
src/obsidian.js and the old utils paths keep re-export shims so no import
site changed; all pre-move tests passed unchanged through the shims before
any test moved.

Test custody (nothing dropped): 2446 baseline -> 2448 total
(+2 new heartbeatPayload writer/reader round-trips). Package side: 74.
App side: 2374 (2446 - 72 moved). Split, not moved wholesale:
deterministicBlockIds (goldens + normalization to the package; transport
round-trips stay), completionMarkers (grammar + parse scoping move; write/
native/detection stay), frontmatter (never-emit rule + parser premise +
helpers move; creation-emission premises stay), heartbeat (shape/staleness
move; transport reads stay), tasksMetadataRead (parse mapping moves;
adoption/hazard/carry policy suites stay), blockIds (helpers + parse
identity + updateTaskLines mechanics move; write round-trips + sync e2e
stay). Moved whole: the tasksMetadata grammar suite. Deliberately NOT
moved: the no-op-write suite — it tests the transports' skip, which stayed;
the byte-determinism it relies on is pinned by the moved grammar suites.

One addition (per instruction, not a rewrite): heartbeatPayload in the
package, now consumed by the plugin in place of its hand-written payload —
the first proof the boundary works in both directions. Verified the
file:-linked package bundles into the plugin's main.js (tsc + esbuild
clean, no external @glance-apps refs in the artifact). Hand-maintained
types/index.d.ts gives the TS plugin a typed surface.

Spec: new 3.11 records the package, the boundary principle, the
file:-now/publish-at-extraction plan, the directory-submission trigger
(root-manifest requirement, verified against the official docs), and the
test-custody rule.

Suite: 173 files / 2448 tests, all green; eslint clean; electron tsc clean.
