+++ package.json
@@ -30,6 +30,7 @@
   "dependencies": {
     "@glance-apps/billing": "0.2.2",
     "@glance-apps/intents": "1.4.0",
+    "@glance-apps/obsidian-format": "file:packages/obsidian-format",
     "@glance-apps/sync": "1.10.0",
     "@modelcontextprotocol/node": "2.0.0",
     "@modelcontextprotocol/server": "2.0.0",
