+++ src/obsidian.js
@@ -21,8 +21,27 @@ import {
 import { unportableEntryReason } from './utils/vaultPortability.js';
 import { blockIdWritesEnabled } from './utils/obsidianWritePolicy.js';
 import { detectTwoSidedRetitle, appendTitleConflictNote, stripObsidianDisplayTag } from './utils/obsidianTitleConflict.js';
-import { splitTasksMetadata, reattachTasksMetadata } from './utils/obsidianTasksMetadata.js';
-import { parseObsidianHeartbeat } from './utils/obsidianHeartbeat.js';
+// The vault line-format core now lives in @glance-apps/obsidian-format
+// (packages/obsidian-format — format, never policy; see its README). This
+// module keeps the TRANSPORTS and the SYNC/MERGE POLICY, imports the format
+// core, and re-exports the names it always exported so no import site or
+// test changed in the extraction.
+import {
+  assertSafeDateStr, formatDatePattern, buildDateParser, parseDateFromFilename, dailyNoteFilename,
+  taskLineSortKey, sortTaskLinesInSection, buildObsidianTaskLine,
+  updateTaskLines, parseTasksFromMarkdown,
+  simpleHash, deriveBlockId, appIdForBlockId, legacyObsidianId,
+  splitBlockId, hasForeignBlockId, blockIdSuffix,
+  splitCompletionMarker, completionMarkerSuffix,
+  splitTasksMetadata, reattachTasksMetadata,
+  parseObsidianHeartbeat,
+} from '@glance-apps/obsidian-format';
+export {
+  formatDatePattern, updateTaskLines, parseTasksFromMarkdown,
+  simpleHash, deriveBlockId, appIdForBlockId, legacyObsidianId,
+  splitBlockId, hasForeignBlockId, blockIdSuffix,
+  splitCompletionMarker, completionMarkerSuffix,
+} from '@glance-apps/obsidian-format';
 import { withCreationFrontmatter } from './utils/obsidianFrontmatter.js';
 
 // How far back the Obsidian daily-note scan reads, in days. DELIBERATELY FIXED and
@@ -235,102 +254,6 @@ async function getDailyNotesDir(vaultHandle, subPath) {
   return current;
 }
 
-/** Reject date strings that aren't strictly YYYY-MM-DD to prevent path injection. */
-function assertSafeDateStr(dateStr) {
-  if (!/^\d{4}-\d{2}-\d{2}$/.test(dateStr)) {
-    throw new Error(`Obsidian: invalid date string "${dateStr}"`);
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Daily note filename pattern helpers
-// Supports Java-style DateTimeFormatter tokens: yyyy, yy, MMMM, MMM, MM, M, dd, d
-// This matches the subset understood by Android's native ObsidianRepository.
-// ---------------------------------------------------------------------------
-
-const DATE_MONTHS_FULL  = ['January','February','March','April','May','June','July','August','September','October','November','December'];
-const DATE_MONTHS_SHORT = DATE_MONTHS_FULL.map(m => m.slice(0, 3));
-const DATE_TOKEN_RE     = /yyyy|yy|MMMM|MMM|MM|M|dd|d/g;
-
-/**
- * Format a Date object into a daily note filename stem using a DateTimeFormatter-style pattern.
- * The default pattern "yyyy-MM-dd" produces "2026-04-05".
- */
-export function formatDatePattern(date, pattern = 'yyyy-MM-dd') {
-  const y = date.getFullYear(), m = date.getMonth() + 1, d = date.getDate();
-  return pattern.replace(DATE_TOKEN_RE, (token) => {
-    switch (token) {
-      case 'yyyy': return y;
-      case 'yy':   return String(y).slice(-2);
-      case 'MMMM': return DATE_MONTHS_FULL[date.getMonth()];
-      case 'MMM':  return DATE_MONTHS_SHORT[date.getMonth()];
-      case 'MM':   return String(m).padStart(2, '0');
-      case 'M':    return m;
-      case 'dd':   return String(d).padStart(2, '0');
-      case 'd':    return d;
-      default:     return token;
-    }
-  });
-}
-
-/**
- * Build a parser object from a pattern string. Used to convert a daily note
- * filename back to a YYYY-MM-DD date string during vault sync.
- */
-function buildDateParser(pattern) {
-  const tokenOrder = [];
-  const captureFor = { yyyy: '(\\d{4})', yy: '(\\d{2})', MMMM: '([A-Za-z]+)', MMM: '([A-Za-z]+)', MM: '(\\d{1,2})', M: '(\\d{1,2})', dd: '(\\d{1,2})', d: '(\\d{1,2})' };
-  let regexStr = '^' + pattern.replace(DATE_TOKEN_RE, (t) => { tokenOrder.push(t); return captureFor[t]; }).replace(/[.+?^${}()|[\]\\]/g, (c) => (tokenOrder.length ? c : '\\' + c)) + '\\.md$';
-  // Rebuild properly: escape literal parts only
-  tokenOrder.length = 0;
-  regexStr = '^';
-  let lastIdx = 0;
-  DATE_TOKEN_RE.lastIndex = 0;
-  let m;
-  while ((m = DATE_TOKEN_RE.exec(pattern)) !== null) {
-    const lit = pattern.slice(lastIdx, m.index);
-    if (lit) regexStr += lit.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-    regexStr += captureFor[m[0]];
-    tokenOrder.push(m[0]);
-    lastIdx = m.index + m[0].length;
-  }
-  const trailing = pattern.slice(lastIdx);
-  if (trailing) regexStr += trailing.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-  regexStr += '\\.md$';
-  return { regex: new RegExp(regexStr, 'i'), tokenOrder };
-}
-
-/**
- * Attempt to parse a daily note filename (e.g. "05-04-2026.md") back to
- * a YYYY-MM-DD string using the given parser. Returns null on failure.
- */
-function parseDateFromFilename(filename, parser) {
-  const match = parser.regex.exec(filename);
-  if (!match) return null;
-  let year, month, day;
-  parser.tokenOrder.forEach((token, i) => {
-    const val = match[i + 1];
-    if (token === 'yyyy')       year  = parseInt(val, 10);
-    else if (token === 'yy')    year  = 2000 + parseInt(val, 10);
-    else if (token === 'MMMM')  month = DATE_MONTHS_FULL.findIndex(n => n.toLowerCase() === val.toLowerCase()) + 1;
-    else if (token === 'MMM')   month = DATE_MONTHS_SHORT.findIndex(n => n.toLowerCase() === val.toLowerCase()) + 1;
-    else if (token === 'MM' || token === 'M') month = parseInt(val, 10);
-    else if (token === 'dd' || token === 'd') day   = parseInt(val, 10);
-  });
-  if (!year || !month || !day) return null;
-  return `${year}-${String(month).padStart(2, '0')}-${String(day).padStart(2, '0')}`;
-}
-
-/**
- * Compute the filename (with .md) for a daily note given an internal YYYY-MM-DD
- * date string and an optional pattern. Returns e.g. "2026-04-05.md".
- */
-function dailyNoteFilename(dateStr, pattern) {
-  if (!pattern || pattern === 'yyyy-MM-dd') return `${dateStr}.md`;
-  const [y, m, d] = dateStr.split('-').map(Number);
-  return `${formatDatePattern(new Date(y, m - 1, d), pattern)}.md`;
-}
-
 /**
  * Write-side file-handle resolution with the portability gate on CREATION
  * ONLY. ORDERING IS THE POINT: the existence check runs BEFORE the validator.
@@ -392,75 +315,6 @@ export async function writeDailyNoteFile(vaultHandle, dailyNotesPath, dateStr, c
 export async function readDailyNoteFresh(vaultHandle, dailyNotesPath, dateStr, pattern) {
   const dirHandle = await getDailyNotesDir(vaultHandle, dailyNotesPath);
   return readDailyNoteFile(dirHandle, dateStr, pattern);
-}
-
-/**
- * Build a sort key for a task line so tasks can be ordered chronologically.
- * Key format: "YYYY-MM-DD THH:MM" — tasks with no date use noteDate (the
- * date of the file), tasks with no time sort after timed tasks of the same date.
- */
-function taskLineSortKey(line, noteDate) {
-  const m = line.match(/^\s*- \[[ xX]\]\s+(.*)$/);
-  if (!m) return '\uffff';
-  const body = m[1].trim();
-  const dateMatch = body.match(/^(\d{4}-\d{2}-\d{2})\s+(.*)$/);
-  const date = dateMatch ? dateMatch[1] : (noteDate || '0000-00-00');
-  const afterDate = dateMatch ? dateMatch[2] : body;
-  const timeMatch = afterDate.match(/^(\d{1,2}):(\d{2})/);
-  if (timeMatch) {
-    return `${date}T${timeMatch[1].padStart(2, '0')}:${timeMatch[2]}`;
-  }
-  return `${date}T\uffff`; // all-day / no-time → sort after timed tasks
-}
-
-/**
- * Sort all top-level task lines within a heading section chronologically,
- * leaving non-task lines (prose, blank lines) after the sorted tasks.
- * Returns a new lines array; the original is not mutated.
- */
-function sortTaskLinesInSection(lines, headingStr, noteDate) {
-  const headingIdx = lines.findIndex(l => l === headingStr);
-  if (headingIdx === -1) return lines;
-  const headingLevel = (headingStr.match(/^#+/) || [''])[0].length;
-  let sectionEnd = lines.length;
-  for (let i = headingIdx + 1; i < lines.length; i++) {
-    const hm = lines[i].match(/^(#+)\s/);
-    if (hm && hm[1].length <= headingLevel) { sectionEnd = i; break; }
-  }
-  const interior = lines.slice(headingIdx + 1, sectionEnd);
-  const taskLines = interior.filter(l => /^\s*- \[/.test(l));
-  const otherLines = interior.filter(l => !/^\s*- \[/.test(l));
-  taskLines.sort((a, b) => {
-    const ka = taskLineSortKey(a, noteDate);
-    const kb = taskLineSortKey(b, noteDate);
-    return ka < kb ? -1 : ka > kb ? 1 : 0;
-  });
-  // Non-task lines (prose, blank lines) go after sorted tasks.
-  // Drop trailing blanks to avoid accumulating empty lines on each write.
-  const nonBlank = otherLines.filter(l => l.trim() !== '');
-  const newSection = [...taskLines, ...nonBlank];
-  if (sectionEnd < lines.length && newSection.length > 0) newSection.push('');
-  return [...lines.slice(0, headingIdx + 1), ...newSection, ...lines.slice(sectionEnd)];
-}
-
-/**
- * Build the formatted markdown task line for a dayGLANCE task, mirroring the
- * format that parseTasksFromMarkdown recognises:
- *   - [ ] Title                         (inbox / all-day today)
- *   - [ ] 2026-03-29 Title              (all-day on another date)
- *   - [ ] 08:00-09:00 Title             (timed task on note's own date)
- *   - [ ] 2026-03-29 08:00-09:00 Title  (timed task on a different date)
- *
- * @param {{ title, startTime, duration, isAllDay, date, blockId }} task
- * @param {string} noteDate  The YYYY-MM-DD date of the note being written to
- */
-function buildObsidianTaskLine(task, noteDate) {
-  const datePrefix = task.date && task.date !== noteDate ? `${task.date} ` : '';
-  const timePrefix = (!task.isAllDay && task.startTime) ? buildTimePrefix(task.startTime, task.duration || null) : '';
-  // Phase 2: a task created in dayGLANCE lands in the vault already carrying
-  // its ^dg- block id, assigned at creation time (useTaskActions) and
-  // persisted on the app task — never derived at read time.
-  return `- [ ] ${datePrefix}${timePrefix}${task.title}${blockIdSuffix(task.blockId, task.title)}`;
 }
 
 /**
@@ -802,133 +656,6 @@ export function detectTasksPluginNative() {
 }
 
 /**
- * Strip leading date / date+time / time prefixes from a raw task line body
- * (the text after `- [x] `) to get the bare title, mirroring
- * parseTasksFromMarkdown.  Returns { bareTitle, datePrefix } where datePrefix
- * is the "YYYY-MM-DD " string (with trailing space) if one was present, or ''.
- */
-function stripLinePrefixes(text) {
-  const trimmed = text.trim();
-  // Regex that matches a single time or a duration range (HH:MM or HH:MM-HH:MM) with optional AM/PM
-  const timeRe = /^(\d{1,2}):(\d{2})\s*(?:[AaPp][Mm])?(?:-\d{1,2}:\d{2}\s*(?:[AaPp][Mm])?)?\s+(.+)$/;
-  // 1) Leading date: "YYYY-MM-DD ..."
-  const dateMatch = trimmed.match(/^(\d{4}-\d{2}-\d{2})\s+(.+)$/);
-  if (dateMatch) {
-    const datePrefix = dateMatch[1] + ' ';
-    const afterDate = dateMatch[2];
-    // Date + time (or date + range)
-    const tm = afterDate.match(timeRe);
-    if (tm) return { bareTitle: tm[3], datePrefix };
-    // Date only
-    return { bareTitle: afterDate, datePrefix };
-  }
-  // 2) Time only (or range only)
-  const tm = trimmed.match(timeRe);
-  if (tm) return { bareTitle: tm[3], datePrefix: '' };
-  // 3) Plain title
-  return { bareTitle: trimmed, datePrefix: '' };
-}
-
-/**
- * Apply a task-state update to daily-note lines IN PLACE. Shared by the FSA
- * and Android-native writeback paths so ID-first matching cannot drift.
- *
- * Matching (Phase 2):
- *  1. ID-first — when the task carries a block id, lines whose trailing
- *     `^dg-<id>` equals it are updated (the id survives the rewrite even when
- *     the title now ends in a user block ref, so identity is never dropped).
- *  2. Fallback — when no line carried the id (line predates tagging, or the
- *     user removed the token), lines are matched by bare title exactly as
- *     before, SKIPPING lines tagged with some other task's id (those are
- *     different tasks now, per the duplicate rule). Fallback-matched lines are
- *     stamped with the task's block id when one is provided — this is the
- *     opportunistic migration moment: existing untagged tasks acquire ids as
- *     they get rewritten, never via a sweep.
- *
- * Updating all occurrences within a match pass mirrors the historical
- * title-dedup behavior.
- *
- * @returns {boolean} whether any line was updated
- */
-export function updateTaskLines(lines, { obsidianRawTitle, completed, startTime, newRawTitle, duration, targetDate, blockId = null, onTitleConflict = null, completedAt = null, completionFormat = null }) {
-  const timeStr = buildTimePrefix(startTime, duration);
-  const writtenTitle = newRawTitle !== undefined ? newRawTitle : obsidianRawTitle;
-  // When targetDate is provided (task rescheduled to a different day), write
-  // an explicit inline date prefix so the task is attributed to the new date
-  // while remaining in its original daily note file.
-  //
-  // The completion marker is a regenerated suffix from task state (see the
-  // marker section): the written title is always stripped of any trailing
-  // marker first — so a hand-written `✅ …` inside a title being stamped for
-  // the first time normalizes out in the same write instead of doubling —
-  // then the marker is re-emitted per (completed, completedAt, format). With
-  // completionFormat null nothing is re-emitted, which is how the OFF
-  // setting converges lines clean on their next touch.
-  const rewrite = (i, indent, datePrefix, idSuffix, title = writtenTitle) => {
-    const effectiveDatePrefix = targetDate ? `${targetDate} ` : datePrefix;
-    // Markers live ONLY on tagged lines (the parse strips them only there).
-    // When this rewrite leaves the line untagged — no id to stamp, or a
-    // foreign block ref refused the stamp — the title stays byte-frozen and
-    // no marker is emitted: a marker the parse can't strip would become
-    // title text and silently change the line's content-derived identity.
-    const tagged = idSuffix !== '';
-    const emitTitle = tagged ? splitCompletionMarker(title).text.trimEnd() : title;
-    const markerSuffix = tagged
-      ? completionMarkerSuffix(completed, completedAt, completionFormat, emitTitle)
-      : '';
-    lines[i] = `${indent}- [${completed ? 'x' : ' '}] ${effectiveDatePrefix}${timeStr}${emitTitle}${markerSuffix}${idSuffix}`;
-  };
-
-  let updated = false;
-  if (blockId) {
-    for (let i = 0; i < lines.length; i++) {
-      const m = lines[i].match(/^(\s*)- \[([ xX])\]\s+(.+)$/);
-      if (!m) continue;
-      const { text: body, blockId: lineId } = splitBlockId(m[3]);
-      if (lineId !== blockId) continue;
-      // Marker-aware splitting: the completion marker is a regenerated
-      // decoration, never title text on a tagged line (the parse strips it
-      // the same way), so the write-time title guard below compares CLEAN
-      // titles — otherwise completing a task would read as a vault-side
-      // retitle.
-      const bodySansMarker = splitCompletionMarker(body).text;
-      const { bareTitle, datePrefix } = stripLinePrefixes(bodySansMarker);
-      // WRITE-TIME TITLE GUARD (the two-sided retitle policy's funnel —
-      // utils/obsidianTitleConflict.js). The line's CURRENT title is the
-      // vault's truth; app state was built from obsidianRawTitle, our last
-      // observation. When the line moved off that base, rebuilding it from
-      // app state would silently revert an Obsidian edit — the one hostile
-      // outcome — so the rewrite KEEPS THE LINE'S OWN TITLE while still
-      // writing the state change. If we were also trying to RETITLE
-      // (newRawTitle differs from the line too), that is a two-sided
-      // conflict: signal it so the caller skips the titleUpdate commit,
-      // obsidianRawTitle stays truthful as the merge base, and the next
-      // scan resolves through the single scan-time policy.
-      const lineDiverged = bareTitle !== obsidianRawTitle && bareTitle !== writtenTitle;
-      if (lineDiverged && newRawTitle !== undefined) onTitleConflict?.({ lineTitle: bareTitle });
-      // Forced suffix: a line that already carried this id keeps it
-      // unconditionally — the foreign-block-ref guard only applies to
-      // FIRST-TIME stamping, never to preserving established identity.
-      rewrite(i, m[1], datePrefix, ` ^dg-${blockId}`, lineDiverged ? bareTitle : writtenTitle);
-      updated = true;
-    }
-    if (updated) return true;
-  }
-
-  for (let i = 0; i < lines.length; i++) {
-    const m = lines[i].match(/^(\s*)- \[([ xX])\]\s+(.+)$/);
-    if (!m) continue;
-    const { text: body, blockId: lineId } = splitBlockId(m[3]);
-    if (lineId) continue; // tagged line — belongs to whichever task owns that id
-    const { bareTitle, datePrefix } = stripLinePrefixes(body);
-    if (bareTitle !== obsidianRawTitle) continue;
-    rewrite(i, m[1], datePrefix, blockIdSuffix(blockId, writtenTitle));
-    updated = true;
-  }
-  return updated;
-}
-
-/**
  * Write a task's completion and scheduling state back to its Obsidian file.
  * Line matching is ID-first with title fallback — see updateTaskLines.
  *
@@ -988,79 +715,6 @@ export async function writeTaskStateToFile(vaultHandle, dailyNotesPath, dateStr,
 // Markdown task parser
 // ---------------------------------------------------------------------------
 
-export function simpleHash(str) {
-  let hash = 0;
-  for (let i = 0; i < str.length; i++) {
-    hash = ((hash << 5) - hash) + str.charCodeAt(i);
-    hash |= 0;
-  }
-  return Math.abs(hash).toString(36);
-}
-
-// ---------------------------------------------------------------------------
-// Block-ID identity (Obsidian build-out Phase 2)
-//
-// Task lines dayGLANCE writes carry a trailing `^dg-<id>` Obsidian block
-// reference — native syntax, invisible in reading view, survives edits and
-// reordering. The id is generated at write time, persisted on the task as
-// `obsidianBlockId`, and embedded in the line itself, so every device that
-// scans the vault derives the same durable identity from the file — unlike
-// the legacy content-derived id (date + title hash), which changes whenever
-// the title or date does. Matching is ID-first with fallback to the legacy
-// text matching for lines that carry no id; existing untagged lines acquire
-// ids opportunistically as they get rewritten. No sweep, no flag day.
-// ---------------------------------------------------------------------------
-
-/**
- * DERIVE a block id: 8 chars of lowercase base36, a pure function of the
- * line's stable identity — (daily-note date, raw title), the same inputs the
- * legacy content id hashes.
- *
- * WHY DERIVED, NOT RANDOM (the echo-stamp decision): with random minting, an
- * edit that reaches the fleet under the LEGACY id (it originated on a device
- * without vault access, or a vault device's write failed before its rename
- * committed) makes every vault-capable device mint its own token for the same
- * line — an N-way identity race that converges only through retirement
- * records, detector reaping, and Obsidian Sync settling the file, with one
- * no-heal corner (a mint whose device closes before ever scanning it lingers
- * as a permanent duplicate row). Deriving the token makes the race
- * semantically empty: every device mints the SAME token, so an N-way mint is
- * N devices writing an identical line — nothing to reconcile, nothing to
- * reap. One logical edit produces one token by UNANIMITY rather than
- * election.
- *
- * NORMALIZATION (pinned deliberately — a subtle cross-device difference here
- * silently restores the old race): the hash input is
- *     `${dateStr}\u0000${String(rawTitle).normalize('NFC').trim()}`
- *   • dateStr — the ISO `YYYY-MM-DD` the write targets (the daily note's
- *     date, exactly the string used in legacy ids and filenames);
- *   • rawTitle — the title AS IT WILL EXIST ON THE LINE (for a retitling
- *     write, the NEW raw title), NFC-normalized (macOS/iOS text input can
- *     produce decomposed forms) and trimmed; internal whitespace is
- *     PRESERVED (titles differing inside are different lines);
- *   • the display '#obsidian' tag never appears in rawTitle by construction;
- *   • NUL separator so ('2026-08-281', 'x') can't alias ('2026-08-28', '1x').
- *
- * ★ FROZEN ALGORITHM: FNV-1a 64-bit over the UTF-8 bytes, mod 36^8, base36,
- * zero-padded to 8. Tokens derived by different app VERSIONS must agree
- * forever — changing any detail here (hash, seed, input shape) reintroduces
- * cross-device divergence between updated and un-updated devices. The golden
- * values in obsidian.deterministicBlockIds.test.js pin it.
- *
- * Collision profile: same inputs as the legacy id, so colliding pairs are
- * exactly the pairs that already collide today, governed by the existing
- * first-occurrence-wins rule; 36^8 ≈ 2.8e12 makes unrelated collisions
- * vanishingly unlikely at vault scale.
- */
-export function deriveBlockId(dateStr, rawTitle) {
-  const input = `${dateStr}\u0000${String(rawTitle ?? '').normalize('NFC').trim()}`;
-  let h = 0xcbf29ce484222325n; // FNV-1a 64-bit offset basis
-  for (const byte of new TextEncoder().encode(input)) {
-    h ^= BigInt(byte);
-    h = (h * 0x100000001b3n) & 0xffffffffffffffffn; // FNV prime, mod 2^64
-  }
-  return (h % 2821109907456n).toString(36).padStart(8, '0'); // 36^8
-}
 
 /**
  * The app-level task id for a block-tagged line. Content-independent, so it is
@@ -1098,394 +752,6 @@ export function buildNewObsidianTaskMeta(rawTitle, todayStr) {
   return { ...base, id: appIdForBlockId(blockId), obsidianBlockId: blockId };
 }
 
-export function appIdForBlockId(blockId) {
-  return `obsidian-dg-${blockId}`;
-}
-
-/** The legacy content-derived id for an untagged line (and pre-Phase-2 tasks). */
-export function legacyObsidianId(taskDate, rawTitle) {
-  return `obsidian-${taskDate}-${simpleHash(rawTitle)}`;
-}
-
-const DG_BLOCK_ID_RE = /\s+\^dg-([a-z0-9]+)\s*$/;
-
-/**
- * Split a trailing ` ^dg-<id>` block reference off a task-line body.
- * Only dayGLANCE's own `^dg-` ids are recognised; a user's own block
- * reference (e.g. `^quote1`) stays part of the title, exactly as today.
- * A `^` anywhere else in the title is untouched.
- * @returns {{ text: string, blockId: string|null }}
- */
-export function splitBlockId(text) {
-  const m = DG_BLOCK_ID_RE.exec(text);
-  if (!m) return { text, blockId: null };
-  return { text: text.slice(0, m.index), blockId: m[1] };
-}
-
-// Obsidian allows exactly one block reference per line, at end of line. A line
-// whose title already ends in a user-authored block id must never get a ^dg-
-// suffix appended — the appended id would become the line's block id and break
-// the user's existing block links.
-const FOREIGN_BLOCK_ID_RE = /(^|\s)\^[A-Za-z0-9-]+$/;
-export function hasForeignBlockId(text) {
-  return FOREIGN_BLOCK_ID_RE.test(String(text).trimEnd());
-}
-
-/**
- * The ` ^dg-<id>` suffix for a written task line, or '' when there is no id to
- * write or the title carries a user-authored block reference (see above).
- */
-export function blockIdSuffix(blockId, writtenTitle) {
-  if (!blockId || hasForeignBlockId(writtenTitle)) return '';
-  return ` ^dg-${blockId}`;
-}
-
-// ---------------------------------------------------------------------------
-// Completion markers (docs/obsidian-buildout-spec.md — completion timestamps)
-//
-// When dayGLANCE completes a task it writes the completion date to the vault
-// line, in whichever of two formats fits the vault: the Tasks plugin's
-// `✅ YYYY-MM-DD` when that plugin is enabled (vault-level detection via
-// .obsidian/community-plugins.json), else a Dataview inline field
-// `[completed:: <ISO date or datetime>]`. The marker is a REGENERATED SUFFIX —
-// like the ^dg- token, it is never part of the stored title on either side —
-// which is what makes re-completion replace rather than append, and
-// uncompletion remove, with no dedicated machinery.
-//
-// IDENTITY SCOPING (the slice of the identity-freeze design this feature
-// needs): the marker is split off at parse time on ^dg--TAGGED LINES ONLY.
-// An untagged line stays byte-frozen exactly as today — a hand-written
-// `✅ 2026-08-10` there remains part of the title. Every line dayGLANCE
-// writes a marker to is tagged (writes stamp), so the strip covers exactly
-// the lines that need it. BOTH formats are always recognised at parse,
-// regardless of current plugin detection — otherwise toggling the Tasks
-// plugin would turn previously-written markers into title text and
-// manufacture title divergence.
-// ---------------------------------------------------------------------------
-
-const TASKS_DONE_MARKER_RE = /\s+✅\s*(\d{4}-\d{2}-\d{2})\s*$/u;
-const DATAVIEW_COMPLETED_RE = /\s+\[completed::\s*([^\]]*?)\s*\]\s*$/;
-
-/**
- * Split a trailing completion marker (either format) off a task-line body
- * that has already had its ^dg- token removed.
- * @returns {{ text: string, completedAt: string|null }} completedAt is the
- *   marker's value when it is ISO-shaped (starts YYYY-MM-DD), else null even
- *   when a marker was stripped.
- */
-export function splitCompletionMarker(text) {
-  let m = TASKS_DONE_MARKER_RE.exec(text);
-  if (m) return { text: text.slice(0, m.index), completedAt: m[1] };
-  m = DATAVIEW_COMPLETED_RE.exec(text);
-  if (m) {
-    return {
-      text: text.slice(0, m.index),
-      completedAt: /^\d{4}-\d{2}-\d{2}/.test(m[1]) ? m[1] : null,
-    };
-  }
-  return { text, completedAt: null };
-}
-
-/**
- * The completion-marker suffix for a written line, or ''.
- *
- * Emission is DETERMINISTIC from the stored completedAt string — the ✅ date
- * is completedAt.slice(0, 10) and the Dataview value is the stored string
- * verbatim — never recomputed from "now" or a local timezone at write time,
- * so every device regenerating the same state emits identical bytes (the
- * no-op write skip depends on this). A date-only historical completedAt
- * emits as a date rather than fabricating a midnight time.
- *
- * `format` null means no marker is written (setting off, or a caller that
- * predates the feature): combined with the unconditional strip in the
- * rewrite paths, turning the setting OFF converges lines clean on their
- * next touch — never via a sweep.
- *
- * Like blockIdSuffix: never append after a user-authored trailing block
- * reference — an Obsidian block id must be the last thing on its line.
- */
-export function completionMarkerSuffix(completed, completedAt, format, writtenTitle) {
-  if (!completed || !completedAt || !format) return '';
-  if (hasForeignBlockId(writtenTitle)) return '';
-  if (format === 'tasks') return ` ✅ ${String(completedAt).slice(0, 10)}`;
-  return ` [completed:: ${completedAt}]`;
-}
-
-/**
- * Try to parse a time string from the beginning of text.
- * Supports single times ("09:00", "9:00 AM") and duration ranges ("09:00-10:00").
- * Returns { startTime, duration, rest } or null.  duration is null when no range.
- */
-function parseLeadingTime(text) {
-  // Try duration range first: HH:MM[-HH:MM] [AM/PM] Title
-  const rangeMatch = text.match(
-    /^(\d{1,2}):(\d{2})\s*([AaPp][Mm])?-(\d{1,2}):(\d{2})\s*([AaPp][Mm])?\s+(.+)$/
-  );
-  if (rangeMatch) {
-    let startH = parseInt(rangeMatch[1], 10);
-    const startM = parseInt(rangeMatch[2], 10);
-    const startAmpm = rangeMatch[3];
-    let endH = parseInt(rangeMatch[4], 10);
-    const endM = parseInt(rangeMatch[5], 10);
-    const endAmpm = rangeMatch[6];
-    if (startAmpm) {
-      const upper = startAmpm.toUpperCase();
-      if (upper === 'PM' && startH < 12) startH += 12;
-      if (upper === 'AM' && startH === 12) startH = 0;
-    }
-    if (endAmpm) {
-      const upper = endAmpm.toUpperCase();
-      if (upper === 'PM' && endH < 12) endH += 12;
-      if (upper === 'AM' && endH === 12) endH = 0;
-    }
-    if (startH < 0 || startH > 23 || endH < 0 || endH > 23) return null;
-    const startTime = `${startH.toString().padStart(2, '0')}:${rangeMatch[2]}`;
-    const rawDuration = (endH * 60 + endM) - (startH * 60 + startM);
-    const duration = rawDuration > 0 ? rawDuration : rawDuration + 1440; // handle midnight wrap
-    return { startTime, duration, rest: rangeMatch[7] };
-  }
-
-  // Fall back to single time: HH:MM [AM/PM] Title
-  const timeMatch = text.match(
-    /^(\d{1,2}):(\d{2})\s*([AaPp][Mm])?\s+(.+)$/
-  );
-  if (!timeMatch) return null;
-  let hours = parseInt(timeMatch[1], 10);
-  const minutes = timeMatch[2];
-  const ampm = timeMatch[3];
-  if (ampm) {
-    const upper = ampm.toUpperCase();
-    if (upper === 'PM' && hours < 12) hours += 12;
-    if (upper === 'AM' && hours === 12) hours = 0;
-  }
-  if (hours < 0 || hours > 23) return null;
-  return {
-    startTime: `${hours.toString().padStart(2, '0')}:${minutes}`,
-    duration: null,
-    rest: timeMatch[4],
-  };
-}
-
-/**
- * Build the time prefix string for writing back to a task line.
- * Produces "HH:MM-HH:MM " when duration is provided, otherwise "HH:MM ".
- */
-function buildTimePrefix(startTime, duration) {
-  if (!startTime) return '';
-  if (!duration) return `${startTime} `;
-  const [h, m] = startTime.split(':').map(Number);
-  const endTotal = h * 60 + m + duration;
-  const eh = Math.floor(endTotal / 60) % 24;
-  const em = endTotal % 60;
-  const endTime = `${eh.toString().padStart(2, '0')}:${em.toString().padStart(2, '0')}`;
-  return `${startTime}-${endTime} `;
-}
-
-/**
- * Parse tasks from Obsidian markdown content.
- *
- * Recognised patterns (in priority order):
- *   - [ ] 2026-02-21 09:00 Date+time task  → scheduled on that date/time
- *   - [ ] 2026-02-21 Date-only task         → all-day task on that date
- *   - [ ] 09:00 Timed task                  → scheduled on the file's date
- *   - [ ] 9:00 AM Timed task                → scheduled on the file's date
- *   - [ ] Simple task                        → inbox task
- *   - [x] Completed task                     → completed (any of the above)
- *
- * Returns { scheduledTasks: [...], inboxTasks: [...] }
- *
- * @param {Set<string>} [seenBlockIds]  duplicate `^dg-` guard. First occurrence
- *   of an id wins; later lines carrying the same id (a copy-paste inside
- *   Obsidian) are parsed as if untagged, becoming ordinary content-derived
- *   tasks. The sync passes ONE set across every file in the scan so the rule
- *   holds vault-wide, not merely per file.
- */
-export function parseTasksFromMarkdown(content, dateStr, seenBlockIds = new Set()) {
-  const scheduled = [];
-  const inbox = [];
-  if (!content) return { scheduledTasks: scheduled, inboxTasks: inbox };
-
-  const lines = content.split('\n');
-
-  for (const line of lines) {
-    // Match: optional whitespace, -, space, [x or space], space, rest
-    const match = line.match(/^\s*- \[([ xX])\]\s+(.+)$/);
-    if (!match) continue;
-
-    const completed = match[1] !== ' ';
-    let rawTitle = match[2].trim();
-
-    // Strip a trailing ^dg-<id> block reference BEFORE any other parsing, so
-    // rawTitle (and therefore the legacy hash) is identical to what an
-    // untagged copy of the line would produce — that identity is what lets
-    // ID-matching and text-matching fall back into each other cleanly.
-    let blockId = null;
-    let lineCompletedAt = null;
-    const idSplit = splitBlockId(rawTitle);
-    if (idSplit.blockId) {
-      rawTitle = idSplit.text.trim();
-      // TAGGED LINES ONLY: split the completion marker out of the title so it
-      // never becomes part of task identity (see the marker section above).
-      // The value is absorbed into completedAt only when the line is actually
-      // checked — a stray marker on an unchecked line is semantically wrong
-      // and is dropped (the next rewrite normalizes the line). Untagged lines
-      // below this branch stay byte-frozen: a hand-written marker there is
-      // title text, exactly as before this feature.
-      const markerSplit = splitCompletionMarker(rawTitle);
-      if (markerSplit.text !== rawTitle) {
-        rawTitle = markerSplit.text.trim();
-        if (completed) lineCompletedAt = markerSplit.completedAt;
-      }
-      if (!seenBlockIds.has(idSplit.blockId)) {
-        seenBlockIds.add(idSplit.blockId);
-        blockId = idSplit.blockId;
-      }
-      // else: duplicate id — first occurrence won; this line falls through as
-      // an untagged task (blockId stays null).
-    }
-
-    let taskDate = dateStr;
-    let startTime = null;
-    let isAllDay = false;
-    let parsedDuration = null;
-
-    // 1) Try inline date: "YYYY-MM-DD ..." at the beginning
-    const dateMatch = rawTitle.match(/^(\d{4}-\d{2}-\d{2})\s+(.+)$/);
-    if (dateMatch) {
-      taskDate = dateMatch[1];
-      const afterDate = dateMatch[2];
-
-      // 1a) Try date + time/range: "YYYY-MM-DD HH:MM[-HH:MM][am/pm] Title"
-      const timePart = parseLeadingTime(afterDate);
-      if (timePart) {
-        startTime = timePart.startTime;
-        if (timePart.duration) parsedDuration = timePart.duration;
-        rawTitle = timePart.rest;
-      } else {
-        // 1b) Date only → all-day task
-        isAllDay = true;
-        rawTitle = afterDate;
-      }
-    } else {
-      // 2) Try time/range only: "HH:MM[-HH:MM][am/pm] Title"
-      const timePart = parseLeadingTime(rawTitle);
-      if (timePart) {
-        startTime = timePart.startTime;
-        if (timePart.duration) parsedDuration = timePart.duration;
-        rawTitle = timePart.rest;
-      }
-    }
-
-    // TASKS METADATA (Phase 4 Step 2, utils/obsidianTasksMetadata.js) —
-    // non-destructive: rawTitle keeps the FULL text (metadata included), so
-    // identity hashing and write matching are untouched on every line,
-    // tagged or not. Extraction derives the display title and maps the
-    // confirmed fields; everything unrecognized stays as preserved text.
-    const meta = splitTasksMetadata(rawTitle);
-    // ⏳ scheduled → the timeline: date-only, so all-day unless the line's
-    // own time prefix supplied a time. An explicit inline date prefix WINS
-    // over ⏳ (it is dayGLANCE's own reschedule channel — when both exist,
-    // the prefix is the newer statement, written by a DG reschedule of a
-    // ⏳-carrying line).
-    if (!dateMatch && meta.fields.scheduled) {
-      taskDate = meta.fields.scheduled;
-      if (!startTime) isAllDay = true;
-    }
-
-    // Add #obsidian tag if not already present — to the DISPLAY title (the
-    // trailing metadata run stripped); rawTitle stays full.
-    const displayBase = meta.text;
-    const title = displayBase.includes('#obsidian') ? displayBase : `${displayBase} #obsidian`;
-
-    // ID-first: a ^dg- tagged line gets its durable block-derived id; an
-    // untagged line keeps the legacy content-derived id (date + title hash).
-    const legacyId = legacyObsidianId(taskDate, rawTitle);
-    const id = blockId ? appIdForBlockId(blockId) : legacyId;
-    // obsidianLegacyId is a PER-SCAN bridge hint, not an identity: it is what
-    // this line's id would have been without the tag, recomputed from current
-    // content each scan. The sync uses it to match a freshly-tagged line to
-    // the task a device still holds under the old id, so app-side fields
-    // survive the one-time legacy → block-id switch.
-    const blockFields = blockId
-      ? { obsidianBlockId: blockId, obsidianLegacyId: legacyId }
-      : {};
-    // Only inject completedAt when a tagged line actually carried a marker —
-    // never an undefined key. The merge rule (app wins; the vault marker
-    // fills a blank) is applied at the sync merge sites, not here.
-    const completedAtFields = lineCompletedAt ? { completedAt: lineCompletedAt } : {};
-    // Mapped Tasks metadata — again only ever present keys, no undefined
-    // injection. 📅 due → deadline (the inbox-deadline concept; a 📅-only
-    // line stays inbox). 🔁 is recognized, never mapped: the flag only
-    // drives the task-card badge saying recurrence is managed in Obsidian.
-    const metadataFields = {
-      ...(meta.fields.due ? { deadline: meta.fields.due } : {}),
-      ...(meta.fields.priority != null ? { priority: meta.fields.priority } : {}),
-      ...(meta.fields.recurrence ? { obsidianRecurrence: true } : {}),
-    };
-
-    if (startTime) {
-      // Timed task (with or without inline date)
-      scheduled.push({
-        id,
-        title,
-        date: taskDate,
-        startTime,
-        duration: parsedDuration || 30,
-        color: 'bg-purple-600',
-        completed,
-        isAllDay: false,
-        notes: '',
-        subtasks: [],
-        importSource: 'obsidian',
-        obsidianRawTitle: rawTitle,
-        obsidianFileDate: dateStr,
-        ...blockFields,
-        ...completedAtFields,
-        ...metadataFields,
-      });
-    } else if (isAllDay) {
-      // Date-only task → all-day scheduled task
-      scheduled.push({
-        id,
-        title,
-        date: taskDate,
-        startTime: '00:00',
-        duration: 30,
-        color: 'bg-purple-600',
-        completed,
-        isAllDay: true,
-        notes: '',
-        subtasks: [],
-        importSource: 'obsidian',
-        obsidianRawTitle: rawTitle,
-        obsidianFileDate: dateStr,
-        ...blockFields,
-        ...completedAtFields,
-        ...metadataFields,
-      });
-    } else {
-      // No date, no time → inbox
-      inbox.push({
-        id,
-        title,
-        priority: 0,
-        completed,
-        notes: '',
-        subtasks: [],
-        duration: 30,
-        color: 'bg-purple-600',
-        importSource: 'obsidian',
-        obsidianRawTitle: rawTitle,
-        obsidianFileDate: dateStr,
-        ...blockFields,
-        ...completedAtFields,
-        ...metadataFields,
-      });
-    }
-  }
-
-  return { scheduledTasks: scheduled, inboxTasks: inbox };
-}
 
 // ---------------------------------------------------------------------------
 // Full vault sync
